backend/extraction_service/requirements.txt
fastapi
uvicorn[standard]
exparso
python-magic
pillow
python-multipart
boto3
python-dotenv
numpy
